+++ lib/reports/index.js
@@ -6,7 +6,7 @@ const { flatten } = require('lodash');
 const getWorkflowStatuses = require('../middleware/get-workflow-statuses');
 
 const pils = require('./pils');
-const pilReviews = require('./pil-reviews');
+const { pilReviews, completed: completedPilReviews } = require('./pil-reviews');
 const namedPeople = require('./named-people');
 const pplList = require('./ppl-list');
 const pplSLA = require('./ppl-sla');
@@ -38,6 +38,7 @@ module.exports = (settings) => {
   const reports = {
     'pils': pils,
     'pil-reviews': pilReviews,
+    'completed-pil-reviews': completedPilReviews,
     'named-people': namedPeople,
     'ppl-list': pplList,
     'ppl-sla': pplSLA,
